Add manager for push's reward
Also tested out the managing of memory via non calling of factory method of create for certain class. It turns out well.
Normally we will utilize calling ::create() of class. But if we have a certain class that manages and consists of multiple classes as a member variable, then
it's better to only call the following in step

1. `Member* obj = new Member();`
2. `obj->init();`
3. Call `CC_SAFE_RELEASE_NULL(obj)` inside dealloc method of that class.

Ensure that `Member` is a derived class of `CCObject` thus as we `new` it. It will have a starting of 1 in reference counting.
When we're done with the enclosing class, we call `release()`, then all of member objects will be released as such.

Note that this is best practice as `CCDirector` class also did this. We try to avoid risking in complex memory releasing here.